# glob matching optimizations