(This is stacked on top of #6, which should land before this one)

This PR adds support for recognizing a number of special cases that come up when using this library inside Stripe.  The general idea is to try and recognize when we can match fixed strings and can therefore avoid doing wildcard matching, which is relatively slow.  There's no better advertisement for this PR than benchmarks.

Before:

```
st-froydnj2:oh-my-glob froydnj$ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/aisamanra/oh-my-glob
BenchmarkCompileLiteralPaths-10                   	 5092194	       226.8 ns/op
BenchmarkLiteralPaths-10                          	16395495	        72.81 ns/op
BenchmarkCompileSubdirFromRoot-10                 	11120170	       107.5 ns/op
BenchmarkSubdirFromRoot-10                        	 7697340	       155.4 ns/op
BenchmarkNegativeSubdirFromRoot-10                	 6224624	       191.6 ns/op
BenchmarkWildcardTail-10                          	20668110	        58.00 ns/op
BenchmarkNegativeWildcardTail-10                  	149957659	         8.088 ns/op
BenchmarkPrefixedSubdirFromRoot-10                	 8335209	       147.8 ns/op
BenchmarkNegativePrefixedSubdirFromRoot-10        	 6492579	       184.7 ns/op
BenchmarkRecursiveFixedFile-10                    	15406861	        77.57 ns/op
BenchmarkNegativeRecursiveFixedFile-10            	17431472	        69.02 ns/op
BenchmarkPrefixedRecursiveFixedFile-10            	15709068	        75.71 ns/op
BenchmarkNegativePrefixedRecursiveFixedFile-10    	17513240	        68.23 ns/op
BenchmarkFixedPathWildcardFile-10                 	13283317	        89.93 ns/op
BenchmarkNegativeFixedPathWildcardFile-10         	 9711602	       125.5 ns/op
PASS
ok  	github.com/aisamanra/oh-my-glob	20.645s
```

After:

```
st-froydnj2:oh-my-glob froydnj$ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/aisamanra/oh-my-glob
BenchmarkCompileLiteralPaths-10                   	 8773447	       129.8 ns/op
BenchmarkLiteralPaths-10                          	24043117	        50.12 ns/op
BenchmarkCompileSubdirFromRoot-10                 	15144303	        78.47 ns/op
BenchmarkSubdirFromRoot-10                        	251199892	         4.753 ns/op
BenchmarkNegativeSubdirFromRoot-10                	269765800	         4.443 ns/op
BenchmarkWildcardTail-10                          	25591287	        46.15 ns/op
BenchmarkNegativeWildcardTail-10                  	98802025	        12.04 ns/op
BenchmarkPrefixedSubdirFromRoot-10                	169671831	         7.058 ns/op
BenchmarkNegativePrefixedSubdirFromRoot-10        	267510488	         4.465 ns/op
BenchmarkRecursiveFixedFile-10                    	236579434	         5.079 ns/op
BenchmarkNegativeRecursiveFixedFile-10            	251166747	         4.824 ns/op
BenchmarkPrefixedRecursiveFixedFile-10            	147275894	         8.171 ns/op
BenchmarkNegativePrefixedRecursiveFixedFile-10    	245319800	         5.128 ns/op
BenchmarkFixedPathWildcardFile-10                 	100000000	        11.22 ns/op
BenchmarkNegativeFixedPathWildcardFile-10         	227851381	         5.576 ns/op
PASS
ok  	github.com/aisamanra/oh-my-glob	23.455s
```

Maybe not the greatest format for eyeballing comparisons, but nearly all the matching benchmarks drop by at least an order of magnitude with some cases closer to 15-20x faster; the `WildcardTail` benchmarks are the exception because I haven't added special cases for them yet.

A slightly earlier version of this branch cuts the time of the globbing step for Stripe's codebase by ~40%.  I'm very excited to see the full version land!